Fixed a few warnings before moving on to realloc.

+++ src/metadata.h
@@ -18,10 +18,10 @@ static MetaData** get_metadata();
 static void* get_data(MetaData* metadata_block);
 
 // Returns the size of the data storded after a metadata block.
-static int get_data_size(MetaData* metadata_block);
+static size_t get_data_size(MetaData* metadata_block);
 
 // Returns the size of a metadata block.
-static int get_metadata_size(MetaData* metadata_block);
+static size_t get_metadata_size(MetaData* metadata_block);
 
-// Allocates room for a metadata and a data at the given adress, and make sure they and the brake are at an adress multiple of 8.
+// Allocates room for a metadata and a data at the given address, and make sure they and the brake are at an address multiple of 8.
 static void* alloc_metadata_and_data(MetaData** metadata_block, size_t size);
